Updates to handle FS and BOTS modes
